# :zap: Create the validator object once