Creating validation object on every request, causes x16 more allocation and performance drops a lot. :zap: This PR fixes it.
Relates: https://github.com/gofiber/fiber/issues/1734

**Before:**
```
Running 20s test @ http://localhost:8080/test
  2 threads and 128 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     7.43ms    7.62ms  86.02ms   86.96%
    Req/Sec    10.52k     3.92k   19.06k    63.50%
  420349 requests in 20.09s, 67.35MB read
Requests/sec:  20923.23
Transfer/sec:      3.35MB
```
![Imgur](https://i.imgur.com/rMVBhoI.png)

After:
```
Running 20s test @ http://localhost:8080/test
  2 threads and 128 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.73ms    1.80ms  54.01ms   90.07%
    Req/Sec    33.06k     8.39k   52.69k    64.75%
  1319443 requests in 20.09s, 211.40MB read
Requests/sec:  65684.51
Transfer/sec:     10.52MB
```
![Imgur](https://i.imgur.com/6VKjWKc.png)